pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chellow"
dynamic = ["version"]
description = "Web Application for checking UK energy bills."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
	"odio==0.0.23",
	"pypdf==6.0.0",
	"pg8000==1.31.1",
	"Flask==3.1.1",
	"SQLAlchemy==2.0.41",
	"flask-restx==1.3.0",
  "jsonschema==4.17.3",  # Temporarily needed until RestX updated
	"openpyxl==3.1.5",
	"python-dateutil==2.8.2",
	"pytz==2022.6",
	"requests==2.32.4",
	"waitress==3.0.1",
	"pep3143daemon==0.1.0",
	"pip>=9.0.1",
	"paramiko==3.4.1",
	"pympler==1.0.1",
	"psutil==5.9.5",
	"xlrd==2.0.1",
	"zish==0.1.12",
]

[project.urls]
"Homepage" = "https://github.com/WessexWater/chellow"

[tool.hatch.version]
path = "VERSION"

[tool.tox]
requires = ["tox>=4.19"]
envlist = ["py"]

[tool.tox.env_run_base]
description = "Run test under {base_python}"
passenv = ["PGPORT", "PGUSER", "PGHOST", "PGPASSWORD"]
allowlist_externals = ["echo", "rm"]
commands = [
    ["black", "--check", "."],
    ["flake8", "."],
    ["validate-pyproject", "pyproject.toml"],
    ["echo", "PGPORT is {env:PGPORT:unset}"],
    ["echo", "PGUSER is {env:PGUSER:unset}"],
    ["echo", "Test database is {env:PGDATABASE:unset} on {env:PGHOST:unset}"],
    ["python", "-m", "pytest", "-v", "-x", "-W error", "test"],
    ["rm", "-rf", "dist"],
    ["python", "-m", "build"],
    ["twine", "check", "dist/*"],
]
deps = [
    "black",
    "build",
    "pytest",
    "pytest-mock",
    "pytest-benchmark",
    "flake8",
    "flake8-alphabetize",
    "Flake8-pyproject",
    "twine",
    "validate-pyproject",
]
[tool.tox.env_run_base.setenv]
PGDATABASE = "toxchellow"

[tool.flake8]
application-names = ["chellow"]
ignore = ["E203", "W503"]
max-line-length = 88
exclude = [".git", "__pycache__", "build", "dist", "venv", ".tox"]
application-import-names = ["chellow"]
